chg: separate masks into different files to better clarity
