+++ pygaming/screen/mask/mask.py
@@ -1,22 +1,12 @@
-"""
-The mask module contains masks. Masks are objects used to select part of an image on which apply an effect.
-There are 5 types of Masks:
-The first type is composed of only one class: MatrixMask
-The second one is composed of binary geometrical masks: Circle, Ellipsis, Rectangle etc.
-The third one is composed of gradient geometrical masks: GradientCircle, GradientRectangle etc.
-The fourth one is composed of masks extracted from arts or from images.
-The last one is combinations or transformation of other masks.
-"""
+"""This mask submodule contains the bases for masks and geometrical masks."""
 
 from abc import ABC, abstractmethod
 from typing import Callable
 import numpy as np
-from PIL import Image
 import cv2
 from pygame import Surface, surfarray as sa, SRCALPHA, draw, Rect
-from ..error import PygamingException
-from ..file import get_file
-from ..settings import Settings
+from ...error import PygamingException
+from ...settings import Settings
 
 # Mask effects
 ALPHA = 'alpha'
@@ -163,6 +153,22 @@ class MatrixMask(Mask):
 
     def _load(self, settings: Settings):
         """Don't do anything as the matrix is already loaded."""
+
+class FreeMask(Mask):
+    """A FreeMask is a mask whose matrix can be freely and explicitly change during the game."""
+
+    def __init__(self, width, height, initial_matrix):
+        super().__init__(width, height)
+        self.__initial_matrix = initial_matrix
+    
+    def _load(self, settings):
+        self.matrix = self.__initial_matrix
+
+    def update_matrix(self, new_matrix: np.ndarray):
+        if self.matrix.shape == new_matrix.shape:
+            self.matrix = new_matrix
+        else:
+            raise ValueError(f"The matrices have two different shapes: {self.matrix.shape}, {new_matrix.shape}")
 
 class Circle(Mask):
     """A Circle is a mask with two values: 0 in the circle and 1 outside."""
@@ -360,330 +366,3 @@ class GradientRectangle(Mask):
         bottom_dist = np.clip((y_indices - self.inner_bottom) / (self.outer_bottom - self.inner_bottom + 1), 0, 1)
 
         self.matrix = self.transition(np.clip(np.sqrt(left_dist**2 + right_dist**2 + top_dist**2 + bottom_dist**2), 0, 1))
-
-class FromArtAlpha(Mask):
-    """A mask from the alpha layer of an art."""
-
-    def __init__(self, art, index: int= 0) -> None:
-
-        super().__init__(art.width, art.height)
-        self.art = art
-        self.index = index
-
-    def _load(self, settings: Settings):
-        need_to_unload = False
-        if not self.art.is_loaded:
-            need_to_unload = True
-            self.art.load(settings)
-
-        self.matrix = 1 - sa.array_alpha(self.art.surfaces[self.index])/255
-
-        if need_to_unload:
-            self.art.unload()
-
-class FromArtColor(Mask):
-    """
-    A mask from a mapping of the color layers.
-    
-    Every pixel of the art is mapped to a value between 0 and 1 with the provided function.
-    Selects only one image of the art based on the index.
-    """
-
-    def __init__(self, art, function: Callable[[int, int, int], float], index: int = 0) -> None:
-        super().__init__(art.width, art.height)
-        self.art = art
-        self.index = index
-        self.map = function
-
-    def _load(self, settings: Settings):
-        need_to_unload = False
-        if not self.art.is_loaded:
-            need_to_unload = True
-            self.art.load(settings)
-
-        self.matrix = np.apply_along_axis(self.map, 2, sa.array2d(self.art.surfaces[self.index]))
-
-        if need_to_unload:
-            self.art.unload()
-
-class FromImageColor(Mask):
-    """
-    A mask from an image.
-    
-    Every pixel of the art is mapped to a value between 0 and 1 with the provided function.
-    """
-
-    def __init__(self, path: str, function: Callable[[int, int, int], float]) -> None:
-        self.path = get_file('images', path)
-        self.im = Image.open(self.path)
-        width, height = self.im.size
-        super().__init__(width, height)
-        self.map = function
-
-    def _load(self, settings: Settings):
-        rgb_array = np.array(self.im.convert('RGB'))
-        self.matrix = np.apply_along_axis(self.map, 2, rgb_array)
-
-class _MaskCombination(Mask, ABC):
-    """MaskCombinations are abstract class for all mask combinations: sum, products and average"""
-
-    def __init__(self, *masks: Mask):
-
-        if any(mask.width != masks[0].width or mask.height != masks[0].height for mask in masks):
-            raise PygamingException("All masks must have the same shape.")
-        super().__init__(masks[0].width, masks[0].height)
-        self.masks = masks
-
-    @abstractmethod
-    def _combine(self, *matrices: np.ndarray) -> np.ndarray:
-        raise NotImplementedError()
-
-    def _load(self, settings: Settings):
-        for mask in self.masks:
-            if not mask.is_loaded():
-                mask.load(settings)
-
-        self._combine(*(mask.matrix for mask in self.masks))
-
-class SumOfMasks(_MaskCombination):
-    """
-    A sum of mask is a mask based on the sum of the matrixes of the masks, clamped between 0 and 1.
-    For binary masks, it acts like union.
-    """
-
-    def _combine(self, *matrices):
-        return np.minimum(np.sum(matrices), 1)
-
-class ProductOfMasks(_MaskCombination):
-    """
-    A product of mask is a mask based on the product of the matrixes of the masks.
-    For binary masks, it acts like intersections.
-    """
-
-    def _combine(self, *matrices):
-        return np.prod(matrices)
-
-class AverageOfMasks(_MaskCombination):
-    """
-    An average of mask is a mask based on the average of the matrixes of the masks.
-    """
-
-    def __init__(self, *masks: Mask, weights= None):
-        if weights is None:
-            weights = [1]*len(masks)
-        super().__init__(*masks)
-        self.weights = weights
-
-    def _combine(self, *matrices):
-        self.matrix = 0
-        for matrix, weight in zip(matrices, self.weights):
-            self.matrix += matrix*weight
-
-        self.matrix /= sum(self.weights)
-
-class BlitMaskOnMask(_MaskCombination):
-    """
-    A blit mask on mask is a mask where the values of the background below (or above) a given threshold are replaced
-    by the values on the foreground.
-    """
-
-    def __init__(self, background: Mask, foreground: Mask, threshold: float = 0, reverse: bool = False):
-        super().__init__(background, foreground)
-        self.threshold = threshold
-        self.reverse = reverse
-    #pylint: disable=arguments-differ
-    def _combine(self, background_matrix, foreground_matrix) -> np.ndarray:
-        self.matrix = background_matrix
-        if self.reverse:
-            positions_to_keep = background_matrix < self.threshold
-        else:
-            positions_to_keep = background_matrix > self.threshold
-        self.matrix[positions_to_keep] = foreground_matrix[positions_to_keep]
-
-class InvertedMask(Mask):
-    """
-    An inverted mask is a mask whose value are the opposite of the parent mask.
-    """
-
-    def __init__(self, mask: Mask):
-        super().__init__(mask.width, mask.height)
-        self._mask = mask
-
-    def _load(self, settings: Settings):
-        if not self._mask.is_loaded():
-            self._mask.load(settings)
-        self.matrix = 1 - self._mask.matrix
-
-class TransformedMask(Mask):
-    """
-    A Transformed mask is a mask whose matrix is the transformation of the matrix of another mask.
-    The transformation must be a numpy vectorized function or a function matrix -> matrix.
-    """
-
-    def __init__(self, mask: Mask, transformation: Callable[[float], float] | Callable[[np.ndarray], np.ndarray]):
-        super().__init__(mask.width, mask.height)
-        self._mask = mask
-        self.transformation = transformation
-
-    def _load(self, settings: Settings):
-        if not self._mask.is_loaded():
-            self._mask.load(settings)
-
-        self.matrix = np.clip(self.transformation(self._mask.matrix), 0, 1)
-        if self.matrix.shape != self._mask.matrix.shape:
-            raise PygamingException(f"Shape of the mask changed from {self._mask.matrix.shape} to {self.matrix.shape}")
-
-class BinaryMask(Mask):
-    """
-    A binary mask is a mask where every values are 0 or 1. It is based on another mask.
-    The matrix of this mask is that every component is 1 if the value on the parent mask
-    is above a thresold and 0 otherwise. (this is reversed if reverse is set to True).
-    """
-
-    def __init__(self, mask: Mask, threshold: float, reverse: bool = False):
-        super().__init__(mask.width, mask.height)
-        self.threshold = threshold
-        self._mask = mask
-        self.reverse = reverse
-
-    def _load(self, settings: Settings):
-        if not self._mask.is_loaded():
-            self._mask.load(settings)
-
-        if self.reverse:
-            positions_to_keep = self._mask.matrix < self.threshold
-        else:
-            positions_to_keep = self._mask.matrix > self.threshold
-
-        self.matrix = np.zeros_like(self._mask.matrix)
-        self.matrix[positions_to_keep] = 1
-
-class FreeMask(Mask):
-    """A FreeMask is a mask whose matrix can be freely and explicitly change during the game."""
-
-    def __init__(self, width, height, initial_matrix):
-        super().__init__(width, height)
-        self.__initial_matrix = initial_matrix
-    
-    def _load(self, settings):
-        self.matrix = self.__initial_matrix
-
-    def update_matrix(self, new_matrix: np.ndarray):
-        if self.matrix.shape == new_matrix.shape:
-            self.matrix = new_matrix
-        else:
-            raise ValueError(f"The matrices have two different shapes: {self.matrix.shape}, {new_matrix.shape}")
-
-class _MovingMask(Mask, ABC):
-    """An abstract class for all the moving masks."""
-    
-    def __init__(self, width, height):
-        super().__init__(width, height)
-        self._velocity_x = 0
-        self._velocity_y = 0
-    
-    def set_velocity(self, vel_x, vel_y):
-        """Update the velocity of the moving mask in pixel/sec. They can be negative"""
-        self._velocity_x = vel_x/1000
-        self._velocity_y = vel_y/1000
-
-    def _get_move(self, loop_duration):
-        return int(self._velocity_x*loop_duration), int(self._velocity_y*loop_duration)
-
-    @abstractmethod
-    def update(self, loop_duration):
-        raise NotImplementedError()
-
-class _WrappingMovingMask(_MovingMask):
-    """An abstract class for all moving masks that would wrap around the edges."""
-
-    def __init__(self, width, height):
-        super().__init__(self, width, height)
-    
-
-    def update(self, loop_duration):
-        """Update the matrix by rearanging the rows and columns."""
-
-        Nx, Ny = self._get_move(loop_duration) 
-
-        # Normalize to a valid range
-        Nx = Nx % self.width if Nx >= 0 else -(abs(Nx) % self.width)
-        Ny = Ny % self.height if Ny >= 0 else -(abs(Ny) % self.height)
-
-        if Nx != 0:
-            self.matrix = np.concatenate((self.matrix[Nx:], self.matrix[:Nx]), axis=0)
-        if Ny != 0:
-            self.matrix = np.concatenate((self.matrix[:, Ny:], self.matrix[:, :Ny]), axis=1)
-
-class _BouncingMovingMask(_MovingMask):
-    """An abstract class for all moving masks that would bounce on the edges."""
-    
-    def __init__(self, width, height, inner_mask: Mask, center: tuple[int, int] = None):
-        super().__init__(width, height)
-        if center is None:
-            center = width//2, height//2
-        self._center = center
-        self._dx, self._dy = inner_mask.width//2, inner_mask.height//2
-        self._inner_mask = inner_mask
-
-    def make_matrix(self):
-        """Create the matrix of the mask based on the inner mask and its position on the screen"""
-        self.matrix = np.eye(self._width, self._height)
-        inner_matrix = self._inner_mask.matrix
-
-        left = self._center[0] - self._dx
-        right = self._center[0] + self._dx
-        top = self._center[1] - self._dy
-        bottom = self._center[1] - self._dy
-
-        start_x = max(0, left)
-        end_x = min(self._width, right + 1)
-        start_y = max(0, top)
-        end_y = min(self._height, bottom + 1)
-
-        inner_start_x = max(0, -left)
-        inner_end_x = self._inner_mask.width - max(0, right + 1 - self._width)
-        inner_start_y = max(0, -top)
-        inner_end_y = self._inner_mask.height - max(0, bottom + 1 - self._height)
-
-        self.matrix[start_y:end_y, start_x:end_x] = inner_matrix[inner_start_y:inner_end_y, inner_start_x:inner_end_x]
-
-    def _load(self, settings):
-        if not self._inner_mask.is_loaded():
-            self._inner_mask.load()
-        self.make_matrix()
-
-    def update(self, loop_duration):
-        """Update the mask."""
-        Nx, Ny = self._get_move(loop_duration)
-        center_x, center_y = self._center[0] + Nx, self._center[1] + Ny
-        # Clip and bounce the position of the center
-        if center_x - self._dx < 0 and self._velocity_x < 0:
-            self._velocity_x *= -1
-            center_x += 2*(self._dx - center_x)
-        elif center_x + self._dx > self.width and self._velocity_x > 0:
-            self._velocity_x *= -1
-            center_x -= 2*(center_x + self._dx - self.width)
-        
-        if center_y - self._dy < 0 and self._velocity_y < 0:
-            self._velocity_y *= -1
-            center_y += 2*(self._dy - center_y)
-        elif center_y + self._dy > self.height and self._velocity_y > 0:
-            self._velocity_y *= -1
-            center_y -= 2*(center_y + self._dy - self.height)
-        
-        self._center = (center_x, center_y)
-        self.make_matrix()
-
-
-class _DisappearingMovingMask(_BouncingMovingMask):
-    """An abstract class for all moving masks that would disappear through on the edges."""
-
-    def __init__(self, width, height):
-        super().__init__(width, height)
-    
-    def update(self, loop_duration):
-        """Update the mask"""
-        Nx, Ny = self._get_move(loop_duration)
-        self._center = self._center[0] + Nx, self._center[1] + Ny   
-        self.make_matrix()
